Integrate argparse + support for interactive use
I wanted to play around with each iteration without having to modify the
code and re-run... so I implemented an `--interactive` flag that can be
passed.

If not passed - the scripts work as they did before by running a query
hard-coded into the script.

If `--interactive` is passed, then the hard-coded example is printed to
the screen, but the user can enter whatever queries they want (and type
`q` or `quit` to stop).

Some scripts also have an `--ollama-model` optional flag that allows the
user to switch which ollama model that's used at runtime (mostly because
my laptop started smoking when trying to run `deepseek-r1:8b` locally).

Links:
------
* [ArgParse Tutorial](https://docs.python.org/3/howto/argparse.html#argparse-tutorial)
* [ArgParse Reference Docs](https://docs.python.org/3/library/argparse.html#module-argparse)